revert versioning and just create new endpoint
