+++ src/test/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerOtherIT.java
@@ -229,4 +229,38 @@ public class LinkedDataControllerOtherIT extends BaseIT {
 				false);
 	}
 
+
+	//Navigation Types Testing
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getNavigationOptionsTest() throws Exception {
+		String actualbody = assertEntity(restTemplate,
+			"/linked-data/wqp/USGS-05427880/navigation/UT?f=json",
+			HttpStatus.OK.value(),
+			null,
+			null,
+			MediaType.APPLICATION_JSON_VALUE,
+			null,
+			true,
+			false);
+		assertThat(new JSONArray(actualbody),
+			sameJSONArrayAs(new JSONArray(getCompareFile(RESULT_FOLDER, "navigation.json"))).allowingAnyArrayOrdering());
+
+	}
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getNavigationOptionsTestBadRequest() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/wqp/USGS-05427880/navigation/XX",
+			HttpStatus.BAD_REQUEST.value(),
+			null,
+			null,
+			null,
+			null,
+			false,
+			false);
+
+	}
+
 }
