--- src/test/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerV2Test.java
@@ -1,111 +1,0 @@
-package gov.usgs.owi.nldi.controllers;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyMap;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doCallRealMethod;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import java.io.UnsupportedEncodingException;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.springframework.http.HttpStatus;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.mock.web.MockHttpServletResponse;
-
-import gov.usgs.owi.nldi.dao.LookupDao;
-import gov.usgs.owi.nldi.dao.StreamingDao;
-import gov.usgs.owi.nldi.services.LogService;
-import gov.usgs.owi.nldi.services.Navigation;
-import gov.usgs.owi.nldi.services.Parameters;
-import gov.usgs.owi.nldi.services.TestConfigurationService;
-
-public class LinkedDataControllerV2Test {
-	@Mock
-	private StreamingDao streamingDao;
-	@Mock
-	private LookupDao lookupDao;
-	@Mock
-	private Navigation navigation;
-	@Mock
-	private Parameters parameters;
-	@Mock
-	private LogService logService;
-
-	private TestConfigurationService configurationService;
-	private LinkedDataControllerV2 controller;
-	private MockHttpServletResponse response;
-	private MockHttpServletRequest request;
-
-	@BeforeEach
-	@SuppressWarnings("unchecked")
-	public void setUp() {
-		MockitoAnnotations.initMocks(this);
-
-		//Need to mock this for only a few tests
-		doCallRealMethod().when(streamingDao).stream(anyString(), anyMap(), any());
-
-		configurationService = new TestConfigurationService();
-		controller = new LinkedDataControllerV2(lookupDao, streamingDao, navigation, parameters, configurationService, logService);
-		response = new MockHttpServletResponse();
-		request = new MockHttpServletRequest();
-
-		when(logService.logRequest(any(HttpServletRequest.class))).thenReturn(BigInteger.ONE);
-		when(lookupDao.getList(any(String.class), anyMap())).thenReturn(new ArrayList<Map<String, Object>>(), null, getTestList());
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	public void getFlowlinesTest() throws Exception {
-		when(lookupDao.getComid(anyString(), anyMap())).thenReturn(null, goodFeature());
-		controller.getFlowlines(request, response, "DoesntMatter", "DoesntMatter", null, null, null, null);
-		verify(logService).logRequest(any(HttpServletRequest.class));
-		verify(logService).logRequestComplete(any(BigInteger.class), any(int.class));
-		//Mock lookupDao 1st response of null means the comid is not found, thus a 404
-		assertEquals(HttpStatus.NOT_FOUND.value(), response.getStatus());
-
-		controller.getFlowlines(request, response, null, null, null, null, null, null);
-		verify(logService, times(2)).logRequest(any(HttpServletRequest.class));
-		verify(logService, times(2)).logRequestComplete(any(BigInteger.class), any(int.class));
-		//Mock lookupDao 2nd response doesn't actually exist, thus causes a 500 when we try to get flowlines
-		assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.value(), response.getStatus());
-	}
-
-	@Test
-	public void getNavigationOptionsTestBad() throws UnsupportedEncodingException {
-		controller.getNavigationOptions(request, response, null, "DoesntMatter", "UT");
-		verify(logService).logRequest(any(HttpServletRequest.class));
-		verify(logService).logRequestComplete(any(BigInteger.class), any(int.class));
-		assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.value(), response.getStatus());
-	}
-
-	public static List<Map<String, Object>> getTestList() {
-		List<Map<String, Object>> rtn = new ArrayList<>();
-		Map<String, Object> entry = new HashMap<>();
-		entry.put("key", "value");
-		rtn.add(entry);
-		return rtn;
-	}
-
-	public static Map<String, Object> goodFeature() {
-		Map<String, Object> rtn = new LinkedHashMap<>();
-		rtn.put(Parameters.COMID, "12345");
-		return rtn;
-	}
-
-
-}
