+++ src/test/java/gov/usgs/owi/nldi/controllers/NetworkControllerFlowlineIT.java
@@ -1,5 +1,6 @@
 package gov.usgs.owi.nldi.controllers;
 
+import org.json.JSONArray;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +17,9 @@ import com.github.springtestdbunit.annotation.DatabaseSetup;
 
 import gov.usgs.owi.nldi.BaseIT;
 import gov.usgs.owi.nldi.transform.FlowLineTransformer;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONArrayAs;
 
 
 @EnableWebMvc
@@ -320,5 +324,310 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 				false,
 				false);
 	}
+
+
+
+	//UT Testing
+	@Test
+	public void getComidUtTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13293474/navigation/UT/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"7",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13293474_UT.json"),
+			true,
+			false);
+	}
+
+	@Test
+	public void getComidUtDistanceTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/UT/flowlines?distance=10",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"9",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13297246_UT_distance_10.json"),
+			true,
+			false);
+	}
+
+
+	@Test
+	public void getComidUtDistanceTestEmptyNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/UT/flowlines?distance=",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"359",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13297246_UT_distance_empty.json"),
+			true,
+			false);
+
+	}
+
+	@Test
+	public void getComidUtDistanceTestAboveMaxNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/UT/flowlines?distance=10000",
+			HttpStatus.BAD_REQUEST.value(),
+			null,
+			null,
+			null,
+			"getNavigationFlowlines.distance: distance must be between 1 and 9999 kilometers",
+			false,
+			false);
+	}
+
+	@Test
+	public void getComidUtDistanceTestBelowMinNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/UT/flowlines?distance=-1",
+			HttpStatus.BAD_REQUEST.value(),
+			null,
+			null,
+			null,
+			"getNavigationFlowlines.distance: distance must be between 1 and 9999 kilometers",
+			false,
+			false);
+	}
+
+	@Test
+	public void getComidUtDiversionTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294158/navigation/UT/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"15",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294158_UT.json"),
+			true,
+			false);
+	}
+
+	//UM Testing
+	@Test
+	public void getComidUmTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13293474/navigation/UM/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"4",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13293474_UM.json"),
+			true,
+			false);
+	}
+
+	@Test
+	public void getComidUmDistanceTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/UM/flowlines?f=json&distance=10",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"6",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13297246_UM_distance_10.json"),
+			true,
+			false);
+	}
+
+	//DM Testing
+	@Test
+	public void getComidDmTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13296790/navigation/DM/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"5",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13296790_DM.json"),
+			true,
+			false);
+	}
+
+	@Test
+	public void getComidDmDiversionsNotIncludedTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294310/navigation/DM/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"42",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294310_DM.json"),
+			true,
+			false);
+	}
+
+	@Test
+	public void getComidDmDistanceTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13293474/navigation/DM/flowlines?distance=10",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"8",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13293474_DM_distance_10.json"),
+			true,
+			false);
+	}
+
+	//DD Testing
+	@Test
+	public void getComidDdTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294310/navigation/DD/flowlines",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"49",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294310_DD.json"),
+			true,
+			false);
+	}
+
+	@Test
+	public void getComidDdDistanceTestNavigation() throws Exception {
+		//We are going to sacrifice a little accuracy for performance, so this does not match the old way...
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294310/navigation/DD/flowlines?distance=11",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"13",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294310_DD_distance_11.json"),
+			true,
+			false);
+	}
+
+	//PP Testing
+	@Test
+	public void getComidPpStopComidInvalidTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297246/navigation/PP/flowlines?stopComid=13297198",
+			HttpStatus.BAD_REQUEST.value(),
+			null,
+			null,
+			MediaType.APPLICATION_JSON_VALUE,
+			getCompareFile(RESULT_FOLDER, "comid_13297246_PP_stop_13297198_navigation.json"),
+			true,
+			true);
+	}
+
+	@Test
+	public void getComidPpStopComidTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297198/navigation/PP/flowlines?stopComid=13297246",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"12",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13297198_PP_stop_13297246.json"),
+			true,
+			false);
+	}
+
+	//Interesting diversion/tributary
+	//There is another simple diversion between 13294248 and 13294242
+	@Test
+	public void interestingTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13293844/navigation/DM/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"7",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13293844_DM_distance_5.json"),
+			true,
+			false);
+
+		assertEntity(restTemplate,
+			"/linked-data/comid/13293844/navigation/DD/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"14",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13293844_DD_distance_5.json"),
+			true,
+			false);
+
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294328/navigation/DM/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"6",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294328_DM_distance_5.json"),
+			true,
+			false);
+
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294328/navigation/DD/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"10",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294328_DD_distance_5.json"),
+			true,
+			false);
+
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294390/navigation/UM/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"6",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294390_UM_distance_5.json"),
+			true,
+			false);
+
+		assertEntity(restTemplate,
+			"/linked-data/comid/13294390/navigation/UT/flowlines?distance=5",
+			HttpStatus.OK.value(),
+			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+			"22",
+			BaseController.MIME_TYPE_GEOJSON,
+			getCompareFile(RESULT_FOLDER, "comid_13294390_UT_distance_5.json"),
+			true,
+			false);
+	}
+
+	//Parameter Error Testing
+	@Test
+	public void badNavigationModeTestNavigation() throws Exception {
+		assertEntity(restTemplate,
+			"/linked-data/comid/13297198/navigation/XX/flowlines",
+			HttpStatus.BAD_REQUEST.value(),
+			null,
+			null,
+			null,
+			"getNavigationFlowlines.navigationMode: must match \"DD|DM|PP|UT|UM\"",
+			false,
+			false);
+	}
+
+	//Navigation Types Testing
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getNavigationOptionsTest() throws Exception {
+		String actualbody = assertEntity(restTemplate,
+			"/linked-data/comid/13294390/navigation/UT?f=json",
+			HttpStatus.OK.value(),
+			null,
+			null,
+			MediaType.APPLICATION_JSON_VALUE,
+			null,
+			true,
+			false);
+		assertThat(new JSONArray(actualbody),
+			sameJSONArrayAs(new JSONArray(getCompareFile(RESULT_FOLDER, "navigation.json"))).allowingAnyArrayOrdering());
+
+	}
+
+
+
 
 }
