+++ src/test/resources/testResult/network/flowline/comid_13297246_PP_stop_13297198_navigation.json
@@ -2,5 +2,5 @@
 	"status": 400,
 	"error": "Bad Request",
 	"message": "{\"errorCode\":310, \"errorMessage\":\"Start ComID must have a hydroseq greater than the hydroseq for stop ComID.\"}",
-	"path": "/nldi/linked-data/v2/comid/13297246/navigate/PP/flowlines"
+	"path": "/nldi/linked-data/comid/13297246/navigation/PP/flowlines"
 }
