+++ src/test/resources/testResult/network/flowline/navigation.json
@@ -1,0 +1,27 @@
+[
+  {
+    "source": "Flowlines",
+    "sourceName": "NHDPlus flowlines",
+    "features": "http://owi-test.usgs.gov:8080/test-url/linked-data/comid/13294390/navigation/UT/flowlines"
+  },
+  {
+    "source": "huc12pp",
+    "sourceName": "huc12pp",
+    "features": "http://owi-test.usgs.gov:8080/test-url/linked-data/comid/13294390/navigation/UT/huc12pp"
+  },
+  {
+    "source": "np21_nwis",
+    "sourceName": "HNDPlusV2_NWIS_Gages",
+    "features": "http://owi-test.usgs.gov:8080/test-url/linked-data/comid/13294390/navigation/UT/np21_nwis"
+  },
+  {
+    "source": "TEST",
+    "sourceName": "TEST Source",
+    "features": "http://owi-test.usgs.gov:8080/test-url/linked-data/comid/13294390/navigation/UT/test"
+  },
+  {
+    "source": "WQP",
+    "sourceName": "Water Quality Portal",
+    "features": "http://owi-test.usgs.gov:8080/test-url/linked-data/comid/13294390/navigation/UT/wqp"
+  }
+]
